# docs/swagger/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  jpcorrect-backend_internal_domain.Event:
    properties:
      activity_id:
        type: string
      actual_duration:
        type: number
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      event_id:
        type: string
      expected_duration:
        type: number
      mode:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.EventMode'
      note:
        type: string
      record_link:
        type: string
      recording_ended_at:
        type: string
      recording_started_at:
        type: string
      recording_started_by:
        type: string
      start_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  jpcorrect-backend_internal_domain.EventMode:
    enum:
    - report
    - conversation
    - review
    type: string
    x-enum-varnames:
    - EventModeReport
    - EventModeConversation
    - EventModeReview
...
  jpcorrect-backend_internal_domain.Mistake:
    properties:
      comment:
        type: string
      confidence:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.MistakeConfidence'
      created_at:
        type: string
      end_offset_sec:
        type: number
      event_id:
        type: string
      fixed_text:
        type: string
      interacted_at:
        type: string
      latest_correction_source:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.MistakeCorrectionSource'
      latest_correction_text:
        type: string
      mistake_id:
        type: string
      note:
        type: string
      origin_text:
        type: string
      start_offset_sec:
        type: number
      type:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.MistakeType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  jpcorrect-backend_internal_domain.MistakeConfidence:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - MistakeConfidenceLow
    - MistakeConfidenceMedium
    - MistakeConfidenceHigh
  jpcorrect-backend_internal_domain.MistakeCorrectionSource:
    enum:
    - initial
    - chatbot
    type: string
    x-enum-varnames:
    - MistakeCorrectionSourceInitial
    - MistakeCorrectionSourceChatbot
  jpcorrect-backend_internal_domain.MistakeType:
    enum:
    - grammar
    - vocab
    - pronunciation
    - advanced
    type: string
    x-enum-varnames:
    - MistakeTypeGrammar
    - MistakeTypeVocab
    - MistakeTypePronunciation
    - MistakeTypeAdvanced
...
host: localhost:8080
info:
  contact: {}
  description: Japanese language correction platform backend API
  title: jpcorrect API
  version: "1.0"
paths:
...
  /v1/mistakes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Mistake data
        in: body
        name: mistake
        required: true
        schema:
          $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a mistake
      tags:
      - mistakes
  /v1/mistakes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a mistake
      tags:
      - mistakes
    get:
      consumes:
      - application/json
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a mistake by ID
      tags:
      - mistakes
    put:
      consumes:
      - application/json
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      - description: Mistake data
        in: body
        name: mistake
        required: true
        schema:
          $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a mistake
      tags:
      - mistakes
  /v1/mistakes/event/{event_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get mistakes by event ID
      tags:
      - mistakes
  /v1/mistakes/user/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jpcorrect-backend_internal_domain.Mistake'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get mistakes by user ID
      tags:
      - mistakes
  /v1/practices:
    post:
      consumes:
      - application/json
      parameters:
      - description: Practice data
        in: body
        name: practice
        required: true
        schema:
          $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a practice
      tags:
      - practices
  /v1/practices/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Practice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a practice
      tags:
      - practices
    get:
      consumes:
      - application/json
      parameters:
      - description: Practice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a practice by ID
      tags:
      - practices
    put:
      consumes:
      - application/json
      parameters:
      - description: Practice ID
        in: path
        name: id
        required: true
        type: string
      - description: Practice data
        in: body
        name: practice
        required: true
        schema:
          $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a practice
      tags:
      - practices
  /v1/practices/user/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get practices by user ID
      tags:
      - practices
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"